Change how metadata is displayed and stored (#211)
The metadata display has always been an uncomfortable solution in ekg.
This change moves it to another type of interaction, one where it's all
done by emacs commands, instead of buffer editing. This allows a few
hacks to disappear and generally, notes to behave better.

In addition, we store information about ekg schema in the schema itself,
so there isn't a need to have multiple alists to change just to add new
properties.